Update main.py
Added process_images function: This function processes all images in the given input directory and saves the output in the given output directory.
Used os module: The os module is used to check if the output directory exists and to create it if it doesn't.
Used os.listdir: The os.listdir function is used to get a list of all files in the input directory.
Checked file extensions: The code checks if each file has a .jpg or .png extension before processing it.
Used os.path.join: The os.path.join function is used to join the directory path with the file name to get the full path to the image file.